Add new endpoint and update version and asset info
Added a new endpoint `addWithFileToDatabase` in `AssignmentController.cs` to handle the uploading of an assignment along with a file to the database. Updated `AssemblyInformationalVersionAttribute` in `WebApplication2.AssemblyInfo.cs` to a new version identifier. Updated hash values in `WebApplication2.AssemblyInfoInputs.cache` and `staticwebassets.build.json` to reflect changes. Added a new asset entry in `staticwebassets.build.json` for a PowerPoint file in `wwwroot/uploads`.